# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "C:\\OSPanel\\domains\\rest_api\\app"
        }
    ],
    "paths": {
        "/api/organization/address": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get list of organizations by address",
                "description": "Returns list of buildings with organizations by address",
                "operationId": "findByAddress",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Address of organzation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "suite"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organization/coordinates": {
            "get": {
                "tags": [
                    "Coordinates"
                ],
                "summary": "Get list of organizations by coordinates",
                "description": "Returns list of buildings with organizations by coordinates",
                "operationId": "findByCoordinates",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "path",
                        "description": "Latitude",
                        "required": true,
                        "schema": {
                            "type": "double"
                        },
                        "example": "1"
                    },
                    {
                        "name": "longitude",
                        "in": "path",
                        "description": "Longitude",
                        "required": true,
                        "schema": {
                            "type": "double"
                        },
                        "example": "1"
                    },
                    {
                        "name": "distance",
                        "in": "path",
                        "description": "Distance",
                        "required": true,
                        "schema": {
                            "type": "double"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organization/building": {
            "get": {
                "tags": [
                    "geo"
                ],
                "summary": "Get list of organizations by building id",
                "description": "Returns list of buildings with organizations by building id",
                "operationId": "findByBuilding",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of building",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organization/business": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get list of organizations by business",
                "description": "Returns list of buildings with organizations by business",
                "operationId": "findByBusiness",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the Business type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Еда"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/organization/id": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get the organization by id",
                "description": "Returns the organization by id",
                "operationId": "findById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Organization",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/organization/name": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get the organization by name",
                "description": "Returns the organization by name",
                "operationId": "findByName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the Organization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "inc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "BaseModel": {
                "properties": {
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "deleted_at": {
                        "description": "Soft delete timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "Building": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "address": {
                        "description": "Building address",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "address_latitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "address_longitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Business"
                }
            },
            "Business": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "Business name",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Business"
                }
            },
            "Organization": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "Organization name",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "phone": {
                        "description": "Organization phones",
                        "type": "string",
                        "readOnly": "true",
                        "example": "main:+797879888"
                    },
                    "building_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Organization"
                }
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "token",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Address",
            "description": "Address"
        },
        {
            "name": "Coordinates",
            "description": "Coordinates"
        },
        {
            "name": "geo",
            "description": "geo"
        }
    ],
    "security": [
        {
            "token": []
        }
    ]
}